🎨✨ Add laps feature + refactoring
